+++ src/library/librarybackend.h
@@ -88,7 +88,7 @@ class LibraryBackendInterface : public QObject {
   // Returns a section of a song with the given filename and beginning. If the section
   // is not present in library, returns invalid song.
   // Using default beginning value is suitable when searching for single-section songs.
-  virtual Song GetSongByFilename(const QString& filename, int beginning = 0) = 0;
+  virtual Song GetSongByFilename(const QString& filename, qint64 beginning = 0) = 0;
 
   virtual void AddDirectory(const QString& path) = 0;
   virtual void RemoveDirectory(const Directory& dir) = 0;
@@ -144,7 +144,7 @@ class LibraryBackend : public LibraryBackendInterface {
                                const QString& column);
 
   SongList GetSongsByFilename(const QString& filename);
-  Song GetSongByFilename(const QString& filename, int beginning = 0);
+  Song GetSongByFilename(const QString& filename, qint64 beginning = 0);
 
   void AddDirectory(const QString& path);
   void RemoveDirectory(const Directory& dir);
